+++ packages/nav-frontend-skjema/src/radio-gruppe.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as classNames from "classnames";
 import { SkjemaGruppe, SkjemaGruppeProps } from ".";
-
+import "nav-frontend-skjema-style";
 class RadioGruppe extends React.Component<SkjemaGruppeProps> {
   render() {
     const { children, className, ...other } = this.props;
